separated drawing of player into separate file
